Add infrastructure for the distributor concept.
The dbreps script acts as a central hub for report generation.  The
name of the report to be generated is given with the "-r" option, and
the site to be worked on with the "-s" option.

Each report is a module beneath "reports/" that implements a class
"report" subclassed from "reports.report".  See
reports/tests/onepage.py and reports/tests/multiplepages.py for
examples.

For each site there must be a section with that name in the
configuration file ~/.dbreps.ini (or any other configuration file that
is given by the "-c" option).

New options in these sections are:

- "host", "host-commons", "host-commons-user", "host-user": These give
  the name of the SQL server to connect to in general or in case the
  report requires a server with user databases or a copy of the
  Commons database.  For example, sensible settings for enwiki may be
  "sql-s1-rr", "sql-s1", "sql-s1-user", and "sql-s1-user".

- "userdb": This is the name of the user database where the reports
  should store and retrieve data.