Add structure for writing database reports in Rust
It uses a `Report` trait, similar to the current Python report
abstraction.

Everything will run through one executable, which simplifies our
crontab handling. Each report has a frequency set, which will get
checked to see if an update is needed. Everything is async so it
should be pretty trivial to run reports in parallel if we need to.

Port the currently broken uncatcats, which needed an update for
the actor migration.