Get rid of the run!() macro
Mostly just added extra complexity for little value because the macro syntax
is more complicated than normal Rust. The main advantage of it was that
we could pass the client and pool as one argument, which is now collected
in the `Runner` struct.

A bit more verbose but hopefully easier to understand and follow.