Hopefully fix "Unused templates (filtered)" report
The where condition of `lt_id IS NULL` appears to be wrong, we actually
want to know whether the templatelinks rows are empty, so we need to
LEFT JOIN against that.

While we're at it, inline all the filtering prefixes and suffixes into
the query, which makes it much easier to debug since the results are now
in the hundreds instead of tens of thousands.